DONE/llm_backend/api/main.py
import time
import logging
from dotenv import load_dotenv

load_dotenv()

from fastapi import FastAPI, Request
from fastapi.middleware.cors import CORSMiddleware
from ..config import load_app_configs
from .dynamic_router import build_router

# Configure basic logging (JSON format recommended for prod)
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger("API")

app = FastAPI(title="LLM Backend")

# 1. Security: CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # Restrict this in production!
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# 2. Observability: Request Logging Middleware
@app.middleware("http")
async def log_requests(request: Request, call_next):
    start_time = time.time()
    
    response = await call_next(request)
    
    process_time = (time.time() - start_time) * 1000
    
    logger.info(
        f"Path: {request.url.path} | Method: {request.method} | "
        f"Status: {response.status_code} | Latency: {process_time:.2f}ms"
    )
    
    return response

configs = load_app_configs("apps.yaml")

for cfg in configs:
    router = build_router(cfg)
    app.include_router(router, prefix=f"/{cfg.app_name}")
