DONE/llm_backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
llama-index-core>=0.10.0
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-vector-stores-postgres
asyncpg>=0.29.0
pgvector>=0.2.0
langgraph>=0.0.10
pydantic>=2.0.0
